fix MacOS Build

+++ ui/typeview.h
@@ -175,7 +175,7 @@ public:
 	virtual std::vector<std::string> getNavigationModes() override;
 
 	uint64_t findMatchingLine(const BinaryNinja::QualifiedName& name, uint64_t offset,
-		uint64_t& cursorOffset);
+		size_t& cursorOffset);
 	bool navigateToType(const std::string& name, uint64_t offset = 0);
 
 	virtual void OnTypeDefined(BinaryNinja::BinaryView* view, const BinaryNinja::QualifiedName& name,
